requirements.txt
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.4.0
sentence-transformers==2.5.0
numpy==1.26.0
scipy==1.12.0
pandas==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0
streamlit==1.31.0
tqdm==4.66.0
pyyaml==6.0.1
python-dotenv==1.0.0
pytest==8.0.0
